fix scheduler initialization for HF trainer (was not considering length of dataloader)
